Get rustc version string at crate build time
Summary:There's no need to fetch the rustc version at proc-macro
invocation time because its always constant for a given compilation of
the crate. Since Rust doesn't have a stable ABI, the compiler used to
compile rustversion is guaranteed to be the same as the one invoking the
rustversion proc macro.